refactor: rename and restructure all files & functions
