+++ packages/blade/src/storybook-recipes/AccessibilityLayer/AccessibilityLayer.d.ts
@@ -1,0 +1,2 @@
+export { default } from './AccessibilityLayer.web';
+export * from './AccessibilityLayer.web';
