+++ packages/blade/src/storybook-recipes/AccessibilityLayer/getCheckboxA11yProps.ts
@@ -1,4 +1,4 @@
-import mapA11yProps from '../../utils/mapA11yProps';
+import makeAccessible from '../../utils/makeAccessible';
 
 type CheckboxA11yProps = { isDisabled?: boolean; isChecked?: boolean };
 
@@ -6,7 +6,7 @@ export const getCheckboxA11yProps = ({
   isDisabled,
   isChecked,
 }: CheckboxA11yProps): Record<string, unknown> => {
-  return mapA11yProps({
+  return makeAccessible({
     accessibilityDisabled: isDisabled,
     accessibilityRole: 'checkbox',
     accessibilityChecked: isChecked,
