+++ packages/blade/src/utils/makeAccessible/index.ts
@@ -1,0 +1,1 @@
+export { default } from './makeAccessible';
