--- packages/blade/src/utils/mapA11yProps/index.ts
@@ -1,1 +1,0 @@
-export { default } from './mapA11yProps';
