--- packages/blade/src/utils/mapA11yProps/mapA11yProps.d.ts
@@ -1,1 +1,0 @@
-export { default } from './mapA11yProps.web';
